Add style reset file and re-organized files
