+++ src/Components/MovieContainer/MovieContainer.scss
