+++ src/Components/MovieDetails/MovieDetails.scss
